Update project structure and add new features
- Reorganized project structure with Utils/ directory
- Updated matching pipeline and scoring algorithms
- Added new data files and configurations
- Moved LinkedIn URL finder to Utils/
- Updated API structure and frontend
- Added Vercel ignore file and configurations
