+++ .gitignore
@@ -10,3 +10,4 @@ __pycache__/
 
 # Outros
 .DS_Store
+JSONs/jobs_teste.json
