+++ .vercelignore
@@ -1,0 +1,11 @@
+__pycache__/
+*.pyc
+*.ipynb
+*.log
+.venv/
+.vscode/
+*.md
+dataExploration.ipynb
+JSONs/jobs_teste.json
+JSONs/vagas_padrao.json
+Utils/
