+++ Utils/linkedin_url_finder/requirements.txt
@@ -1,0 +1,4 @@
+pandas>=1.3.0
+selenium>=4.0.0
+beautifulsoup4>=4.9.3
+requests>=2.25.1
