+++ api.py
@@ -1,0 +1,88 @@
+from fastapi import Form
+from fastapi import FastAPI, UploadFile, File
+from fastapi.responses import JSONResponse
+from fastapi.middleware.cors import CORSMiddleware
+import shutil, os, json
+from Matching.preparingJobs import load_and_filter_jobs, transform_jobs
+from Matching.pipeline import match_jobs_candidates
+
+app = FastAPI()
+
+# Add CORS middleware to allow frontend requests
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],  # In production, specify your domain
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+@app.get("/health")
+def health_check():
+    return {"status": "healthy", "message": "API de matching ativa"}
+
+@app.post("/match_vaga")
+async def match_vaga_text(descricao: str = Form(...)):
+    # 1. Monta objeto de vaga temporário
+    vaga = {"id": "vaga_unica", "descricao": descricao}
+
+    # 2. Carrega candidatos
+    candidates_path = os.path.join(os.path.dirname(__file__), "JSONs", "candidates.json")
+    if not os.path.exists(candidates_path):
+        return JSONResponse({"erro": "Arquivo de candidatos não encontrado."}, status_code=400)
+    with open(candidates_path, "r", encoding="utf-8") as f:
+        candidates = json.load(f)
+
+    # 3. Aplica o matching
+    res = match_jobs_candidates([vaga], candidates)
+
+    # 4. Monta resposta: top 3 candidatos para a vaga
+    match = res["top_matches"][0]
+    top_candidatos = [
+        {"candidato": c["cand_id"], "score": c["match_score"]}
+        for c in match["top"]
+    ]
+    return {"vaga": descricao, "top_candidatos": top_candidatos}
+
+@app.post("/match_vagas")
+async def match_vagas(file: UploadFile = File(...)):
+    # 1. Recebe o JSON e salva como vagas.json
+    vagas_path = "/tmp/vagas.json"
+    with open(vagas_path, "wb") as buffer:
+        shutil.copyfileobj(file.file, buffer)
+
+    # 2. Faz as transformações corretas no JSON
+    filtered_jobs = load_and_filter_jobs()
+    if not filtered_jobs:
+        return JSONResponse({"erro": "Erro ao carregar ou filtrar vagas."}, status_code=400)
+    jobs_list = transform_jobs(filtered_jobs)
+
+    # 3. Carrega candidatos
+    candidates_path = os.path.join(os.path.dirname(__file__), "JSONs", "candidates.json")
+    if not os.path.exists(candidates_path):
+        return JSONResponse({"erro": "Arquivo de candidatos não encontrado."}, status_code=400)
+    with open(candidates_path, "r", encoding="utf-8") as f:
+        candidates = json.load(f)
+
+
+    # 4. Aplica o matching
+    res = match_jobs_candidates(jobs_list, candidates)
+
+    # 5. Monta resposta: top 3 candidatos para cada vaga
+    top_matches = []
+    for match in res["top_matches"]:
+        top_matches.append({
+            "vaga": match["job_id"],
+            "top_candidatos": [
+                {"candidato": c["cand_id"], "score": c["match_score"]}
+                for c in match["top"]
+            ]
+        })
+
+    # 6. Apaga o arquivo temporário de vagas
+    if os.path.exists(vagas_path):
+        os.remove(vagas_path)
+
+    return {"top_matches": top_matches}
+
+app_handler = app  # Para garantir compatibilidade com Vercel
