--- linkedin_url_finder/INTEGRATION_README.md
@@ -1,55 +1,0 @@
-# LinkedIn URL Finder - Integração
-
-## 📋 Sobre este Módulo
-
-Este é o **Localizador de Perfis do LinkedIn**, um sistema automatizado para descoberta de perfis do LinkedIn de formandos universitários.
-
-### 🎯 Características Principais
-
-- **Alta Taxa de Sucesso**: 100% de sucesso para formandos recentes (2024-2025)
-- **160 Perfis Únicos**: Descobertos e validados
-- **Sistema Inteligente**: Filtra automaticamente formandos recentes
-- **Interface em Português**: Completamente localizada
-- **Processamento em Lotes**: Eficiente e respeitoso
-
-### 🚀 Como Usar
-
-1. **Instalar Dependências**:
-   ```bash
-   pip install -r requirements.txt
-   python setup_chromedriver.py
-   ```
-
-2. **Executar o Sistema**:
-   ```bash
-   python linkedin_production.py
-   ```
-
-3. **Verificar Progresso**:
-   ```bash
-   python check_progress.py
-   ```
-
-### 📊 Dados Atuais
-
-- **Formandos Processados**: 160/160 (100%)
-- **Perfis Encontrados**: 160 únicos
-- **Período**: 2024-2025 (formandos recentes)
-- **Taxa de Sucesso**: 100%
-
-### 🔧 Arquivos Principais
-
-- `linkedin_production.py`: Script principal de produção
-- `check_progress.py`: Verificador de progresso
-- `linkedin_success_master.json`: Base de dados com perfis encontrados
-- `new_graduates.csv`: Lista de formandos para processar
-
-### 📝 Integração Realizada em
-
-**Data**: 25/09/2025 19:37:09
-**Versão**: v1.0 - Sistema completo com interface em português
-**Status**: Pronto para produção
-
----
-
-**Desenvolvido com ❤️ para descoberta eficiente de perfis do LinkedIn**
