+++ vercel.json
@@ -1,24 +1,15 @@
 {
   "version": 2,
   "builds": [
-    {
-      "src": "api/index.py",
-      "use": "@vercel/python"
-    }
+    { "src": "api.py", "use": "@vercel/python" }
   ],
   "routes": [
-    {
-      "src": "/api/(.*)",
-      "dest": "/api/index.py"
-    },
-    {
-      "src": "/(.*)",
-      "dest": "/public/$1"
-    }
-  ],
-  "functions": {
-    "api/index.py": {
-      "runtime": "python3.9"
-    }
-  }
+    { "src": "/", "dest": "/index.html" },
+    { "src": "/health", "dest": "api.py" },
+    { "src": "/match_vaga", "dest": "api.py" },
+    { "src": "/match_vagas", "dest": "api.py" },
+    { "src": "/docs", "dest": "api.py" },
+    { "src": "/openapi.json", "dest": "api.py" },
+    { "src": "/(.*)", "dest": "/index.html" }
+  ]
 }
